select option and search for real weather

+++ internal/geolocation/geolocation.go
@@ -4,8 +4,8 @@ import "fmt"
 
 type GeolocationResult struct {
 	Name    string  `json:"name"`
-	Lat     float32 `json:"lat"`
-	Lng     float32 `json:"lon"`
+	Lat     float64 `json:"lat"`
+	Lng     float64 `json:"lon"`
 	Country string  `json:"country"`
 	State   string  `json:"state"`
 }
